+++ types/Media/Alignment.ts
@@ -1,0 +1,6 @@
+enum Alignment {
+  CENTRE = 'centre',
+  JUSTIFIED = 'justified',
+}
+
+export default Alignment;
