+++ types/Media/Orientation.ts
@@ -1,0 +1,6 @@
+enum Orientation {
+  LANDSCAPE = 'landscape',
+  PORTRAIT = 'portrait',
+}
+
+export default Orientation;
